Split OverlayRenderer logic to OverlayManager
-  Split logic related to populating overlays to OverlayManager and
OverlayRenderer.
- Remove overlay-related logic from plugin, this is going to be handled
via OverlayManager.
- Inject core overlays to OverlayManager via new methods
