+++ runelite-client/src/main/java/net/runelite/client/ui/overlay/OverlayPriority.java
@@ -29,5 +29,6 @@ public enum OverlayPriority
 	LOW,
 	NONE,
 	MED,
-	HIGH
+	HIGH,
+	HIGHEST
 }
